move filter_suggestion to annif.suggestion where it fits better than in annif.util
