+++ tests/test_util.py
@@ -1,7 +1,5 @@
 """Unit tests for Annif utility functions"""
 
-from scipy.sparse import csr_array
-
 import annif.util
 
 
@@ -19,21 +17,3 @@ def test_metric_code():
 
     for input, output in zip(inputs, outputs):
         assert annif.util.metric_code(input) == output
-
-
-def test_filter_suggestion_limit():
-    pred = csr_array([[0, 1, 3, 2], [1, 4, 3, 0]])
-    filtered = annif.util.filter_suggestion(pred, limit=2)
-    assert filtered.toarray().tolist() == [[0, 0, 3, 2], [0, 4, 3, 0]]
-
-
-def test_filter_suggestion_threshold():
-    pred = csr_array([[0, 1, 3, 2], [1, 4, 3, 0]])
-    filtered = annif.util.filter_suggestion(pred, threshold=2)
-    assert filtered.toarray().tolist() == [[0, 0, 3, 2], [0, 4, 3, 0]]
-
-
-def test_filter_suggestion_limit_and_threshold():
-    pred = csr_array([[0, 1, 3, 2], [1, 4, 3, 0]])
-    filtered = annif.util.filter_suggestion(pred, limit=2, threshold=3)
-    assert filtered.toarray().tolist() == [[0, 0, 3, 0], [0, 4, 3, 0]]
